requirements.txt
fastapi
fastapi_login
bcrypt
pydantic
uvicorn
configparser
pymysql
python-multipart
cryptography
sqlalchemy
python-dotenv


